refactor: entities/tag 제거
