+++ frontend/techpick/src/apis/tag/deleteTag.ts
@@ -1,5 +1,5 @@
 import { apiClient } from '@/apis';
-import { TagType } from '../type';
+import { TagType } from '@/types';
 
 export const deleteTag = async (deleteTagId: TagType['tagId']) => {
   await apiClient.delete<never>(`tag/${deleteTagId}`);
