+++ frontend/techpick/src/components/SelectedTagItem/SelectedTagCommonStyle.css.ts
