+++ frontend/techpick/src/components/SelectedTagItem/SelectedTagContent.css.ts
