+++ frontend/techpick/src/components/SelectedTagItem/SelectedTagContent.tsx
