+++ frontend/techpick/src/components/SelectedTagItem/SelectedTagLayout.css.ts
