+++ frontend/techpick/src/components/SelectedTagItem/SelectedTagLayout.tsx
