+++ frontend/techpick/src/components/SelectedTagItem/index.ts
@@ -1,3 +1,2 @@
 export { SelectedTagItem } from './SelectedTagItem';
-export { SelectedTagListLayout } from './SelectedTagListLayout';
 export { SelectedTagCommonStyle } from './SelectedTagCommonStyle.css';
