+++ frontend/techpick/src/components/SelectedTagListLayout/SelectedTagListLayout.tsx
