+++ frontend/techpick/src/components/index.ts
@@ -4,3 +4,5 @@ export * from './Gap';
 export { DeferredComponent } from './DeferredComponent';
 export { PickCard } from './PickCard/PickCard';
 export { PickCardGridLayout } from './PickCardGridLayout/PickCardGridLayout';
+export { SelectedTagItem } from './SelectedTagItem';
+export { SelectedTagListLayout } from './SelectedTagListLayout/SelectedTagListLayout';
