--- frontend/techpick/src/entities/tag/index.ts
@@ -1,4 +1,0 @@
-export type * as tagTypes from './type';
-export * from './ui/SelectedTagItem';
-export * from './model/tagStore';
-export * from './api';
