+++ frontend/techpick/src/types/index.ts
@@ -1,1 +1,2 @@
 export type { NodeData, DirectoryNodeProps } from './NodeData';
+export type * from './tag.type';
