+++ frontend/techpick/src/types/tag.type.ts
